[UFO-Race] fix minimap offset

+++ libraries/Gamebuino-Meta/examples/6.Games/UFO-Race/src/UFO-Race/world.ino
@@ -101,9 +101,9 @@ void drawWorld(){
 void drawMap(){
   gb.display.fillRect(0,0,18,18);
   gb.display.setColor(WHITE);
-  int8_t tile_x = (player.x/16) - 8;
+  int8_t tile_x = (((uint16_t)player.x)/16) - 8;
   int8_t tile_x_initial = tile_x;
-  int8_t tile_y = (player.y/16) - 8;
+  int8_t tile_y = (((uint16_t)player.y)/16) - 8;
   for(byte y = 0; y < 16; y++, tile_y++){
     for(byte x = 0; x < 16; x++, tile_x++){
       if(tile_x < 0 || tile_x > WORLD_W || tile_y < 0 || tile_y > WORLD_H)
